Simplify the stats for now

+++ app/models/updaters/pvp/base_stats.rb
@@ -5,16 +5,18 @@ class Updaters::Pvp::BaseStats
 
   def update
     @users.each do |user|
-      user.pvp_base_stat.destroy if user.pvp_base_stat
-      ::PvpBaseStat.create({
-                            user: user,
-                            kills: @stats[user.id]['mergedAllCharacters']['results']['allPvP']['allTime']['kills']['basic']['value'],
-                            deaths: @stats[user.id]['mergedAllCharacters']['results']['allPvP']['allTime']['deaths']['basic']['value'],
-                            kd: @stats[user.id]['mergedAllCharacters']['results']['allPvP']['allTime']['killsDeathsRatio']['basic']['value'],
-                            suicides: @stats[user.id]['mergedAllCharacters']['results']['allPvP']['allTime']['suicides']['basic']['value'],
-                            assists: @stats[user.id]['mergedAllCharacters']['results']['allPvP']['allTime']['assists']['basic']['value'],
-                            most_kills: @stats[user.id]['mergedAllCharacters']['results']['allPvP']['allTime']['bestSingleGameKills']['basic']['value']
-                          })
+      unless @stats[user.id].nil?
+        user.pvp_base_stat.destroy if user.pvp_base_stat
+        ::PvpBaseStat.create({
+                               user: user,
+                               kills: @stats[user.id]['Response']['mergedAllCharacters']['results']['allPvP']['allTime']['kills']['basic']['value'],
+                               deaths: @stats[user.id]['Response']['mergedAllCharacters']['results']['allPvP']['allTime']['deaths']['basic']['value'],
+                               kd: @stats[user.id]['Response']['mergedAllCharacters']['results']['allPvP']['allTime']['killsDeathsRatio']['basic']['value'],
+                               suicides: @stats[user.id]['Response']['mergedAllCharacters']['results']['allPvP']['allTime']['suicides']['basic']['value'],
+                               assists: @stats[user.id]['Response']['mergedAllCharacters']['results']['allPvP']['allTime']['assists']['basic']['value'],
+                               most_kills: @stats[user.id]['Response']['mergedAllCharacters']['results']['allPvP']['allTime']['bestSingleGameKills']['basic']['value']
+                             })
+      end
     end
   end
 end
